refactor config for .env
